feat: polish release
