+++ .changeset/violet-dogs-design.md
@@ -1,0 +1,5 @@
+---
+'@coinbase/onchainkit': patch
+---
+
+- **feat**: added `getTokens`, which helps retrieve a list of tokens on Base. By @0xAlec #389
