+++ site/docs/pages/token/get-tokens.mdx
@@ -1,0 +1,37 @@
+# `getTokens`
+
+The `getTokens` utility is designed to retrieve a list of tokens on Base.
+
+## Usage
+
+:::code-group
+
+```tsx [code]
+import { getTokens } from '@coinbase/onchainkit/token';
+
+const tokens = await getTokens({ limit: '1', search: 'degen' });
+```
+
+```ts [return value]
+[
+  {
+    address: '0x4ed4e862860bed51a9570b96d89af5e1b0efefed',
+    chainId: 8453,
+    decimals: 18,
+    image:
+      'https://d3r81g40ycuhqg.cloudfront.net/wallet/wais/3b/bf/3bbf118b5e6dc2f9e7fc607a6e7526647b4ba8f0bea87125f971446d57b296d2-MDNmNjY0MmEtNGFiZi00N2I0LWIwMTItMDUyMzg2ZDZhMWNm',
+    name: 'DEGEN',
+    symbol: 'DEGEN',
+  },
+];
+```
+
+:::
+
+## Returns
+
+[`Promise<GetTokensResponse>`](/token/types#gettokensresponse)
+
+## Parameters
+
+[`GetTokensOptions`](/identity/types#gettokensoptions)
