+++ site/docs/pages/token/types.mdx
@@ -5,6 +5,31 @@ description: Glossary of Types in Token Kit.
 
 # Types [Glossary of Types in Token Kit.]
 
+## `GetTokensError`
+
+```ts
+type GetTokensError = {
+  code: number; // The Error code
+  error: string; // The Error message
+};
+```
+
+## `GetTokensResponse`
+
+```ts
+type GetTokensResponse = Token[] | GetTokensError;
+```
+
+## `GetTokensOptions`
+
+```ts
+type GetTokensOptions = {
+  limit?: string; // The maximum number of tokens to return (default: 50)
+  search?: string; // A string to search for in the token name or symbol
+  page?: string; // The page number to return (default: 1)
+};
+```
+
 ## `Token`
 
 ```ts
