+++ src/token/types.ts
@@ -19,6 +19,28 @@ export type LegacyTokenData = {
 /**
  * Note: exported as public Type
  */
+export type GetTokensError = {
+  code: number; // The Error code
+  error: string; // The Error message
+};
+
+/**
+ * Note: exported as public Type
+ */
+export type GetTokensResponse = Token[] | GetTokensError;
+
+/**
+ * Note: exported as public Type
+ */
+export type GetTokensOptions = {
+  limit?: string; // The maximum number of tokens to return (default: 50)
+  search?: string; // A string to search for in the token name or symbol
+  page?: string; // The page number to return (default: 1)
+};
+
+/**
+ * Note: exported as public Type
+ */
 export type Token = {
   address: Address; // The address of the token contract
   chainId: number; // The chain id of the token contract
@@ -34,25 +56,3 @@ export type Token = {
 export type TokenChipReact = {
   token: Token;
 };
-
-/**
- * Note: exported as public Type
- */
-export type GetTokensOptions = {
-  limit?: string;
-  search?: string;
-  page?: string;
-};
-
-/**
- * Note: exported as public Type
- */
-export type GetTokensError = {
-  code: number;
-  error: string;
-};
-
-/**
- * Note: exported as public Type
- */
-export type GetTokensResponse = Token[] | GetTokensError;
