+++ src/version.ts
@@ -1,1 +1,1 @@
-export const version = '0.17.1';
+export const version = '0.17.2';
